+++ packages/design-system-web/.storybook/preview.js
@@ -14,11 +14,8 @@ export const parameters = {
     storySort: {
       order: [
         'Welcome',
-        'Design',
-        'Advanced Components',
-        'Basic Components',
-        'Grid Components',
-        'Contributing',
+        'Design', ['Overview', 'Layout', 'Typography', 'Theming', 'Theme Generator', 'Theme Generator Color'],
+        'Contributing'
       ],
     },
   },