+++ packages/design-system-web/.storybook/react/component.ts
@@ -3,12 +3,10 @@
 /* auto-generated react proxies */
 import { createReactComponent } from './react-component-lib';
 
-import type { JSX } from '@geovistory/design-system-web';
+import type { JSX } from '../../dist/types';
 
-import { defineCustomElements } from '@geovistory/design-system-web/loader';
 
-defineCustomElements();
-export const GvHelloworld = /*@__PURE__*/createReactComponent<JSX.GvHelloworld, HTMLGvHelloworldElement>('gv-helloworld');
+
 export const IonAccordion = /*@__PURE__*/createReactComponent<JSX.IonAccordion, HTMLIonAccordionElement>('ion-accordion');
 export const IonAccordionGroup = /*@__PURE__*/createReactComponent<JSX.IonAccordionGroup, HTMLIonAccordionGroupElement>('ion-accordion-group');
 export const IonActionSheet = /*@__PURE__*/createReactComponent<JSX.IonActionSheet, HTMLIonActionSheetElement>('ion-action-sheet');