+++ packages/design-system-angular/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.3.1-pr-14.8 (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 0.3.1-pr-14.7 (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 0.3.1-pr-14.6 (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 0.3.1-pr-14.5 (2022-07-06)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 0.3.1-pr-14.4 (2022-07-06)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 0.3.1-pr-14.3 (2022-07-06)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 0.3.1-pr-14.2 (2022-07-05)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 0.3.1-pr-14.1 (2022-07-01)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 0.3.1-pr-14.0 (2022-06-28)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
 ## 0.3.0 (2022-06-17)
 
 **Note:** Version bump only for package @geovistory/design-system-angular